Authorization Users

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -4,12 +4,13 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CategoryRequest;
+use App\Models\Category;
 use App\Repositories\Category\CategoryInterface;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Request;
 use Illuminate\View\View;
 use Throwable;
 
@@ -34,9 +35,12 @@ class CategoryController extends Controller
      * Display a listing of the resource.
      *
      * @return Application|Factory|View
+     * @throws AuthorizationException
      */
     public function list()
     {
+        $this->authorize('view', Category::class);
+
         $categories = $this->category->all(['childs']);
 
         return view('admin.categories.list', compact('categories'));
@@ -46,9 +50,12 @@ class CategoryController extends Controller
      * New Category Form.
      *
      * @return JsonResponse
+     * @throws AuthorizationException|Throwable
      */
     public function form()
     {
+        $this->authorize('create', Category::class);
+
         $categories = $this->category->all(['childs']);
 
         if (!$categories) {
@@ -72,6 +79,8 @@ class CategoryController extends Controller
      */
     public function store(CategoryRequest $request)
     {
+        $this->authorize('create', Category::class);
+
         $newCategory = $this->category->new($request->only('name', 'parent_id', 'description'), true);
 
         if (!$newCategory) {
@@ -96,11 +105,14 @@ class CategoryController extends Controller
     /**
      * Display specific category.
      *
-     * @param Request $request
+     * @param $id
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function show($id)
     {
+        $this->authorize('view', Category::class);
+
         $category = $this->category->find($id);
 
         $parentCategories = $this->category->findAll(null, '=', 'parent_id');
@@ -124,9 +136,13 @@ class CategoryController extends Controller
      * @param CategoryRequest $request
      * @param $id
      * @return JsonResponse
+     * @throws AuthorizationException
+     * @throws Throwable
      */
     public function update(CategoryRequest $request, $id)
     {
+        $this->authorize('update', Category::class);
+
         $update = $this->category->edit($id, $request->only('name', 'parent_id', 'description'));
 
         $categories = $this->category->all(['childs']);
@@ -142,9 +158,13 @@ class CategoryController extends Controller
      *
      * @param $id
      * @return JsonResponse
+     * @throws AuthorizationException
+     * @throws Throwable
      */
     public function destroy($id)
     {
+        $this->authorize('delete', Category::class);
+
         $remove = $this->category->destroy($id);
 
         $categories = $this->category->all(['childs']);
